refactor: split reducer into two reducers
